feat: ROS2 Humble package, Docker, CI, and comprehensive README
ROS2 workspace (ros2_ws/):
- adaptive_pid_msgs (ament_cmake): PIDGains, PlantState, TrainingStats messages
- adaptive_pid_ros2 (ament_python): 7-node system, all reusing the same
  InvertedPendulumPlant / PIDController / GainScheduler / DomainRandomizer /
  DisturbanceObserver classes verified in tests/ -- plant_node, reference_node,
  pid_controller_node, rl_agent_node, disturbance_node, logging_node,
  visualization_node
- Launch file bringing up the full node graph with PPO/SAC selectable via arg
- All node files verified to compile (py_compile) -- rclpy itself requires a
  ROS2 Humble environment (provided via Docker) to actually run
- ros2_ws/README.md with build/run instructions

Docker & CI:
- docker/Dockerfile: ROS2 Humble base + full Python RL stack, builds ros2_ws
- docker/docker-compose.yaml, .dockerignore
- .github/workflows/ci.yml: lint (ruff/mypy), unit tests (matrix 3.10/3.11),
  simulation+integration tests, training smoke tests, Docker build

Documentation:
- Comprehensive top-level README: motivation, architecture, MDP summary,
  installation, training, evaluation, ROS2 deployment, results status,
  engineering-discipline/verification section, honest limitations, future work

Code quality pass:
- Fixed all ruff lint findings (redundant int(round()), mutable class
  attribute needing ClassVar, unused unpacked variables, dict() literals)
- Full suite re-verified: 115/115 tests passing, ruff clean, ROS2 files
  syntax-clean

This completes the initial full-repository build described in the project
brief: architecture docs, control layer, MuJoCo simulation, reward function,
domain randomization, Gymnasium environment, PPO/SAC training, evaluation/
benchmark suite with publication-quality plotting, ROS2 Humble deployment,
Docker, and CI, all built and verified incrementally module-by-module.

+++ scripts/autotune_zn.py
@@ -35,7 +35,7 @@ def simulate_proportional_step_response(
     callback ``find_ultimate_gain`` needs (see control/ziegler_nichols.py).
     """
     plant.reset(initial_theta=initial_theta, initial_theta_dot=0.0)
-    n_steps = int(round(duration_s / dt))
+    n_steps = round(duration_s / dt)
     thetas = np.zeros(n_steps)
     for i in range(n_steps):
         state = plant.get_state()
